refactor: LV3 객체지향 설계 완료 ver.1.0 - 공통된 Input 구현 클래스들 삭제 - 함수, 변수명 변경
