+++ README.md
@@ -1,1 +1,25 @@
-# sparta-java-assignment-calculator
+# 계산기 과제
+
+---
+
+## 요구 사항
+
+### LV 1
+
+- 양의 정수 입력 받기
+- 사칙 연산 기호 입력 받기
+- 3개의 입력으로 출력
+
+### LV 2
+
+- 연산 결과 값을 저장하는 컬렉션
+- Lv 1의 Calculator 클래스의 활용
+- 컬렉션 필드의 캡슐화
+- 가장 먼저 저장된 데이터의 삭제 구현
+
+### LV 3
+
+- 제네릭을 사용하여 입력받은 피연산자의 타입
+- Enum을 활용한 연산 타입 정의
+- 저장된 데이터 중 보다 큰 결과값들 출력
+
