+++ diagram.md
@@ -1,0 +1,10 @@
+---
+title: Bank example
+---
+classDiagram
+class BankAccount
+BankAccount : +String owner
+BankAccount : +Bigdecimal balance
+BankAccount : +deposit(amount)
+BankAccount : +withdrawal(amount)
+
