+++ src/lv03/io/input/CalculatorInput.java
@@ -1,17 +1,14 @@
-package lv03.mainMenu.calculationMenu.io.input;
-
-import lv03.input.Input;
+package lv03.io.input;
 
 import java.util.Scanner;
 
-public class CalculationMenuInput implements Input {
+public class CalculatorInput implements Input {
     private final Scanner scanner;
 
-    public CalculationMenuInput(Scanner scanner) {
+    public CalculatorInput(Scanner scanner) {
         this.scanner = scanner;
     }
 
-    @Override
     public String input() {
         System.out.print(">> ");
         return scanner.nextLine();
