--- src/lv03/mainMenu/io/output/mainMenuOutput.java
@@ -1,4 +1,0 @@
-package lv03.mainMenu.io.output;
-
-public class mainMenuOutput {
-}
